% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateMetropolisHastingsincpp.R
\name{cppmultivariatemh}
\alias{cppmultivariatemh}
\title{Sample data from a multivariate distribution using the Metropolis-Hastings Algorithm using a cpp implementation.
The use of this function is recommended when the target density function is in the predefined set of densities in the function's API.}
\usage{
cppmultivariatemh(
  targetdensity = c("Exponential"),
  candidatedensity = c("Normal"),
  numIter = 1000,
  initial_vec = NULL,
  plot = FALSE,
  sigma_matrix = NULL,
  seed = 1001L
)
}
\arguments{
\item{targetdensity}{This is the probability density function from which sampling will be done. This should be one of the supported values:
"Exponential"- Exponential distribution.}

\item{candidatedensity}{This is the candidate generation density to be used. This should be one of the supported values}

\item{numIter}{The number of samples to generate}

\item{initial_vec}{The staring vector for the algorithm}

\item{plot}{plot the values of x}

\item{sigma_matrix}{standard deviation matrix of the candidate generation density}

\item{seed}{seed for generation of random samples}
}
\value{
a matrix of sample vectors
}
\description{
Sample data from a multivariate distribution using the Metropolis-Hastings Algorithm using a cpp implementation.
The use of this function is recommended when the target density function is in the predefined set of densities in the function's API.
}
