% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivariateMetropolisHastingsincpp.r
\name{cppunivariatemh}
\alias{cppunivariatemh}
\title{Sample data from a univariate distribution using the Metropolis-Hastings Algorithm using a cpp implementation.}
\usage{
cppunivariatemh(
  targetdensity = c("Exponential"),
  candidatedensity = c("Normal"),
  numIter = 1000,
  initial = 0,
  plot = FALSE,
  sigma = 1,
  seed = 1001L
)
}
\arguments{
\item{targetdensity}{This is the probability density function from which sampling will be done. This should be a function of x.}

\item{candidatedensity}{This is the candidate generation density to be used}

\item{numIter}{The number of samples to generate}

\item{initial}{The staring value for the algorithm}

\item{plot}{plot the values of x}

\item{sigma}{standard deviation of the candidate generation density}

\item{seed}{seed for generation of random samples}
}
\value{
a vector of samples
}
\description{
Sample data from a univariate distribution using the Metropolis-Hastings Algorithm using a cpp implementation.
}
