% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multivariatemhexpnormalcpp}
\alias{multivariatemhexpnormalcpp}
\title{C++ function to generate samples using multivariate normal candidate density}
\usage{
multivariatemhexpnormalcpp(numIter, initial_vec, sigma_matrix)
}
\arguments{
\item{numIter}{This is the number of iterations to run the algorithm}

\item{initial_vec}{This is the starting vector for the markov chain}

\item{sigma_matrix}{This is the standard deviation matrix for candidate generating multivariate normal distribution.}
}
\description{
C++ function to generate samples using multivariate normal candidate density
}
