% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateMetropolisHastings.r
\name{rmultivariatemh}
\alias{rmultivariatemh}
\title{Sample data from a univariate distribution using the Metropolis-Hastings Algorithm}
\usage{
rmultivariatemh(
  targetdensity,
  candidatedensity = c("Normal"),
  numIter = 1000,
  initial_vec = NULL,
  plot = FALSE,
  sigma_matrix = NULL,
  seed = 1001L
)
}
\arguments{
\item{targetdensity}{This is the probability density function from which sampling will be done. This should be a function of x.}

\item{candidatedensity}{This is the candidate generation density to be used}

\item{numIter}{The number of samples to generate}

\item{initial_vec}{The staring vector for the algorithm}

\item{plot}{plot the values of x}

\item{sigma_matrix}{standard deviation matrix of the candidate generation density}

\item{seed}{seed for generation of random samples}
}
\value{
a matrix of sample vectors
}
\description{
Sample data from a univariate distribution using the Metropolis-Hastings Algorithm
}
