% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{targetDensityMultivarExp}
\alias{targetDensityMultivarExp}
\title{C++ function to generate a probability from a multivariate exponential density function}
\usage{
targetDensityMultivarExp(x)
}
\arguments{
\item{x}{This is the x value}
}
\description{
C++ function to generate a probability from a multivariate exponential density function
}
