% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{targetDensityUnivarExp}
\alias{targetDensityUnivarExp}
\title{C++ function to generate a probability from a univariate exponential density function}
\usage{
targetDensityUnivarExp(x)
}
\arguments{
\item{x}{This is the x value}
}
\description{
C++ function to generate a probability from a univariate exponential density function
}
