% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{univariatemhexpnormalcpp}
\alias{univariatemhexpnormalcpp}
\title{C++ function to generate samples using univariate normal candidate density}
\usage{
univariatemhexpnormalcpp(numIter, initial, sigma)
}
\arguments{
\item{numIter}{This is the number of iterations to run the algorithm}

\item{initial}{This is the starting value for the markov chain}

\item{sigma}{This is the standard deviation for candidate generating univariate normal distribution.}
}
\description{
C++ function to generate samples using univariate normal candidate density
}
