% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{univariatemhnormalcpp}
\alias{univariatemhnormalcpp}
\title{C++ function to generate samples using univariate normal candidate density}
\usage{
univariatemhnormalcpp(targetDensity, numIter, initial, sigma)
}
\arguments{
\item{targetDensity}{This is the probability density function from which sampling will be done. This should be a function of x.}

\item{numIter}{This is the number of iterations to run the algorithm}

\item{initial}{This is the starting value for the markov chain}

\item{sigma}{This is the standard deviation for candidate generating univariate normal distribution.}
}
\description{
C++ function to generate samples using univariate normal candidate density
}
